# riscv: add crc32 optimization using zbc extension