--- a/arch/riscv/riscv_functions.h
+++ b/arch/riscv/riscv_functions.h
@@ -22,6 +22,10 @@ void slide_hash_rvv(deflate_state *s);
 void inflate_fast_rvv(PREFIX3(stream) *strm, uint32_t start);
 #endif
 
+#ifdef RISCV_CRC32_ZBC
+uint32_t crc32_riscv64_zbc(uint32_t crc, const uint8_t *buf, size_t len);
+#endif
+
 #ifdef DISABLE_RUNTIME_CPU_DETECTION
 // RISCV - RVV
 #  if defined(RISCV_RVV) && defined(__riscv_v) && defined(__linux__)
@@ -43,6 +47,12 @@ void inflate_fast_rvv(PREFIX3(stream) *strm, uint32_t start);
 #    define native_longest_match_slow longest_match_slow_rvv
 #    undef native_slide_hash
 #    define native_slide_hash slide_hash_rvv
+#  endif
+
+// RISCV - CRC32
+#  if (defined(RISCV_CRC32_ZBC) && defined (__riscv_zbc))
+#    undef native_crc32
+#    define native_crc32 crc32_riscv64_zbc
 #  endif
 #endif