Benchmark tested in musebook, which supports RISCV Zbc Extension, 4x~7x speedup when data_len >= 16.
```
...
71: crc32/generic_chorba/1                          17.5 ns         17.5 ns     39945854
71: crc32/generic_chorba/8                          51.9 ns         51.9 ns     13179168
71: crc32/generic_chorba/12                         85.7 ns         85.7 ns      8168409
71: crc32/generic_chorba/16                         98.3 ns         98.3 ns      7108298
71: crc32/generic_chorba/32                          191 ns          191 ns      3653600
71: crc32/generic_chorba/64                          308 ns          308 ns      2266995
71: crc32/generic_chorba/512                        1050 ns         1050 ns       665992
71: crc32/generic_chorba/4096                       6422 ns         6422 ns       109009
71: crc32/generic_chorba/32768                     56719 ns        56719 ns        12029
71: crc32/generic_chorba/262144                   395788 ns       395785 ns         1768
71: crc32/generic_chorba/4194304                 8142963 ns      8142532 ns           86
71: crc32/braid/1                                   16.2 ns         16.2 ns     43098429
71: crc32/braid/8                                   50.0 ns         50.0 ns     10000000
71: crc32/braid/12                                  73.6 ns         73.6 ns      9486070
71: crc32/braid/16                                  86.4 ns         86.4 ns      8102579
71: crc32/braid/32                                   179 ns          179 ns      3910695
71: crc32/braid/64                                   263 ns          263 ns      2662938
71: crc32/braid/512                                 1503 ns         1503 ns       465720
71: crc32/braid/4096                               11298 ns        11298 ns        61939
71: crc32/braid/32768                              93140 ns        93140 ns         7530
71: crc32/braid/262144                            740223 ns       740220 ns          945
71: crc32/braid/4194304                         11938022 ns     11938037 ns           59
71: crc32/riscv/1                                   24.2 ns         24.2 ns     29427622
71: crc32/riscv/8                                   57.3 ns         57.3 ns     12435003
71: crc32/riscv/12                                  89.6 ns         89.6 ns      7814444
71: crc32/riscv/16                                   24.7 ns         24.7 ns     28316264
71: crc32/riscv/32                                  30.7 ns         30.7 ns     22812923
71: crc32/riscv/64                                  40.3 ns         40.3 ns     17372502
71: crc32/riscv/512                                  181 ns          181 ns      3883264
71: crc32/riscv/4096                                1306 ns         1306 ns       534089
71: crc32/riscv/32768                              10479 ns        10479 ns        67008
71: crc32/riscv/262144                             82844 ns        82844 ns         8445
71: crc32/riscv/4194304                          1368543 ns      1368549 ns          511
...
```

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

## Summary by CodeRabbit

- **New Features**
  - Added support for RISC-V ZBC hardware acceleration, providing an optimized CRC32 implementation for compatible CPUs.
  - Introduced automatic detection and use of the RISC-V ZBC extension during runtime and build configuration.
- **Bug Fixes**
  - Improved feature detection logic for RISC-V CPU capabilities.
- **Tests**
  - Added new tests and benchmarks for the RISC-V ZBC CRC32 implementation to ensure correctness and performance.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->